What a clusterf

Conflicts:
	Android.mk
	CleanSpec.mk
	include/speex/speex.h
	include/speex/speex_bits.h
	include/speex/speex_buffer.h
	include/speex/speex_callbacks.h
	include/speex/speex_config_types.h
	include/speex/speex_echo.h
	include/speex/speex_header.h
	include/speex/speex_jitter.h
	include/speex/speex_preprocess.h
	include/speex/speex_resampler.h
	include/speex/speex_stereo.h
	include/speex/speex_types.h
	libspeex/arch.h
	libspeex/bits.c
	libspeex/buffer.c
	libspeex/cb_search.c
	libspeex/cb_search.h
	libspeex/cb_search_bfin.h
	libspeex/fftwrap.c
	libspeex/filters.c
	libspeex/filters.h
	libspeex/filters_bfin.h
	libspeex/fixed_bfin.h
	libspeex/jitter.c
	libspeex/lpc.c
	libspeex/lpc_bfin.h
	libspeex/lsp.c
	libspeex/lsp.h
	libspeex/lsp_bfin.h
	libspeex/ltp.c
	libspeex/ltp.h
	libspeex/ltp_bfin.h
	libspeex/mdf.c
	libspeex/misc_bfin.h
	libspeex/modes.c
	libspeex/modes.h
	libspeex/modes_wb.c
	libspeex/nb_celp.c
	libspeex/nb_celp.h
	libspeex/preprocess.c
	libspeex/quant_lsp.c
	libspeex/quant_lsp.h
	libspeex/quant_lsp_bfin.h
	libspeex/resample.c
	libspeex/resample_neon.h
	libspeex/resample_sse.h
	libspeex/sb_celp.c
	libspeex/sb_celp.h
	libspeex/scal.c
	libspeex/speex.c
	libspeex/speex_callbacks.c
	libspeex/speex_header.c
	libspeex/stereo.c
	libspeex/vbr.c
	libspeex/vbr.h
	libspeex/vq.c
	libspeex/vq_bfin.h
	libspeex/window.c

+++ include/speex/speex_config_types.h
@@ -3,9 +3,9 @@
 
 /* these are filled in by configure */
 typedef short spx_int16_t;
-typedef unsigned spx_uint16_t;
+typedef unsigned short spx_uint16_t;
 typedef int spx_int32_t;
-typedef unsigned spx_uint32_t;
+typedef unsigned int spx_uint32_t;
 
 #endif
